airflowApp/requirements.txt
alembic==1.6.5
anyio==3.1.0
apache-airflow==2.1.0
apache-airflow-providers-ftp==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-postgres==1.0.2
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
blinker==1.4
Brotli==1.0.9
cattrs==1.7.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.13
cryptography==3.4.7
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.3.0
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-Compress==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
graphviz==0.16
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
numpy==1.20.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pendulum==2.1.2
prison==0.1.3
psutil==5.8.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-daemon==2.3.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
retrying==1.3.3
rfc3986==1.5.0
rich==10.3.0
setproctitle==1.2.2
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.7
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
unicodecsv==0.14.1
urllib3==1.26.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
